// src/game/game.schema.ts
import { Entity, PBMaterial_PbrMaterial } from "@dcl/sdk/ecs"

export enum E_Letter {
    A = 0, B = 1, C = 2, D = 3, E = 4, F = 5, G = 6, H = 7, I = 8, J = 9, K = 10,
    L = 11, M = 12, N = 13, O = 14, P = 15, Q = 16, R = 17, S = 18, T = 19, U = 20,
    V = 21, W = 22, X = 23, Y = 24, Z = 25
}

export type PbrMat = {
    $case: "pbr";
    pbr: PBMaterial_PbrMaterial;
}

export type LetterIndex = {
    x: number, y: number
}

export enum HighlightColor {
    Yellow = 0,
    Green = 1,
    Red = 2,
    Orange = 3,
    Blue = 4,
    Purple = 5,
}

export type HighlightData = {
    isStart: boolean,
    direction: number,
    scale: number
}

export type HighlightPiece = {
    entity: Entity,
    index: number,
    inUse: boolean
}